Add files via upload
This is my final project while attending the Full-Stack Data Analytics program. The purpose of this project is to find the right method to increase the CRR (Customer Retention Rate) of members in the loyalty program at Canada Airline by 5%, from 68% to 73%, within one year. The actionable insights provided are personalized for the two clusters found: Frequent Long-Distance Travelers and Low-Engagement Customers.